pyproject.toml
[tool.poetry]
name = "omnidb"
version = "0.1.0"
description = ""
authors = ["shabbywu <shabbywu@qq.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <3.10"
django = "^3.1.7"
cherrypy = "^18.6.0"
social-auth-app-django = "^4.0.0"
pyscrypt = "^1.6.2"
pyaes = "^1.6.1"
openpyxl = "^3.0.7"
sshtunnel = "^0.4.0"
psycopg2-binary = "^2.8.6"
restrictedpython = "^5.1"
cx-oracle = "^8.1.0"
pymysql = "^1.0.2"
sqlparse = "^0.4.1"
django-sass = "^1.0.1"
pgspecial = "^1.12.1"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
python-multipart = "^0.0.5"
pyjwt = "^2.0.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
url = "https://mirrors.tencent.com/pypi/simple/"
name = "pypi-tencent-mirror"
default = true
